Check if transaction failed for some packages
When yum is processing transaction and some failure occures (e.g. prein
script fails for package being installed) yum API does not raise an
exception. Instead, it sets output_state of corresponding transaction
member to TS_FAILED and continues.

With this commit we ensure that such cases are handled correctly. In
particular, when it happens:
- action_type in output is set for 'error' for such packages;
- yyoom prints a message to stderr and exits with non-zero exit status.

Co-authored-by: Joshua Harlow<harlowja@yahoo-inc.com>
Change-Id: Idc3823a4dfa3641122b7037fa3c6ba23b869d72f